enhance kg phase 5
