+++ config/json_configs/rag_inference_local.json
@@ -67,7 +67,8 @@
                 "max_facts_after_reranking": 5,
                 "enable_llm_reranking": true,
                 "damping_factor": 0.5,
-                "ppr_backend": "push"
+                "ppr_backend": "push",
+                "ppr_push_epsilon": 0.000001
             }
         ],
         "fusion_method": "rrf",
