+++ core/deepsearch/tools/hybrid/__init__.py
@@ -2,14 +2,12 @@
 
 from .hybrid_probe import HybridNeighborhoodProbeTool
 from .context_rollup import ContextRollupTool
-from .multi_adapter_compare import MultiAdapterComparatorTool
 from .evidence_crosscheck import EvidenceCrosscheckTool
 from .code_python import CodePythonTool
 
 __all__ = [
     "HybridNeighborhoodProbeTool",
     "ContextRollupTool",
-    "MultiAdapterComparatorTool",
     "EvidenceCrosscheckTool",
     "CodePythonTool",
 ]
